Fix downloads badge
In a recent update, pepy.tech has updated the API calls to require API keys. This breaks our "Total Downloads" badge on the README.md.

This PR proposed a temporary fix to restore the badge. 

In this fix, we can only show a "Monthly Downloads" badge instead of "Total Downloads". 

According to the maintainer, they will work with shields.io to make it possible to display "Total Downloads" again.

https://github.com/psincraian/pepy/issues/477#issuecomment-1788812705